Get rid of warnings, incomplete
Recompile binary, now loads without VC redistributable package

+++ g_save.c
@@ -249,7 +249,7 @@ void InitGame (void)
 void WriteField1 (FILE *f, field_t *field, byte *base)
 {
 	void		*p;
-	int			len;
+	size_t		len;
 	int			index;
 
 	p = (void *)(base + field->ofs);
@@ -268,7 +268,7 @@ void WriteField1 (FILE *f, field_t *field, byte *base)
 			len = strlen(*(char **)p) + 1;
 		else
 			len = 0;
-		*(int *)p = len;
+		*(int *)p = (int)len;
 		break;
 	case F_EDICT:
 		if ( *(edict_t **)p == NULL)
@@ -299,7 +299,7 @@ void WriteField1 (FILE *f, field_t *field, byte *base)
 
 void WriteField2 (FILE *f, field_t *field, byte *base)
 {
-	int			len;
+	size_t		len;
 	void		*p;
 
 	p = (void *)(base + field->ofs);
